backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
motor
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
python-multipart